requirements.txt
behave == 1.2.6
configparser == 3.7.4
pyhamcrest == 1.9.0
selenium == 3.141.0
behave-jenkins == 1.2.5.1